Removed OMDb from the project
This also prevents the deprecated routes from creating anything in the database, they can only read existing information that was written by a newer client.
It also stops us serving the "poster" and "ratings" fields for media. The poster has been unused in UMS for a while now.
The ratings is a shame, but TMDB doesn't give that information. However since the user has already decided to get the content, it's probably not that useful to show it in this context.